Fix templated splitStringWithEncs.
Thew problem was with the additional type charType - if I wanted a
default value for that argument, then the compiler couldn't work out
what type that should be.

Solution was to use the string member value_type which is char for a
string and presumably Glib::unichar for a Glib::ustring.