Report PID as loaded for all states except zombie or dead
Use of the pidLoaded() method to stop a cdlrawin process from
continuing if there was already an existing process on the same queue
was giving the incorrect result when the existing process dropped into
the 'disk sleep' state and pidLoaded() returned false. Modify the
behaviour so that a PID is considered to be loaded unless it is found
to be in either zombie or dead state.
(cherry picked from commit 8ca76c15ee04f34a20b9b475c7fc526f5c8a3f9e)